创建排队活动+排号实验+叫号优化+订阅消息

+++ src/main/java/com/vx/vo/ActivityVO.java
@@ -29,5 +29,5 @@ public class ActivityVO {
 
     private String pictureUrl;
 
-    private Float distance;
+    private Double distance;
 }
